backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
jinja2
aiosmtplib
email-validator
python-dotenv
